Install minikube on Windows Server 2019 Hyper-V
If you are looking for a great way to get up to speed with learning
Kubernetes, getting your hands in a lab environment is one of the
best ways to do that. Kubernetes is arguably not for the faint of
heart to deploy correctly and get a working environment.
However, there is a way to easily get up and running with a one
node Kubernetes “cluster” to get a feel for working with
Kubernetes. The solution is called minikube. This post will take a
look at how to install minikube on Windows Server 2019 Hyper-V,
including configuration, interacting with, and deploying
applications. Let’s take a look.